refactor: 파일 구조 변경
- 컴포넌트별 파일 분리
- eventlistener 파일 생성
